slight reorg. comments.
